requirements.txt
fastapi
uvicorn
pandas
scikit-learn
psycopg2-binary
python-dotenv
mlflow
torch --index-url https://download.pytorch.org/whl/cpu